deprecate global variables used in test specs
This was long overdue. This module was created from a test suite. WHen I
extracted the tester from the test suite, I could not figure a way to
specify the test loaded via a "do" command without global variables. Or
maybe I was lazy or tired at that time.

Even in tests, global variables are a bad idea. I got bitten more than
once.

Long story short, this commit deprecate the global variables. The tests
now should be specified with a data structure. This enable user to
enable strictures and warnings in test files. THis offer more ways to
specify the data structure.

On the downside, the user must make sure that the last statement of the
test file returns the data structure. Either with

  return \%test_data;

or

  \%test_data;